+++ packages/web-sdk/src/index.ts
@@ -15,6 +15,7 @@ export {
   ViewInstrumentation,
   WebVitalsInstrumentation,
   SessionInstrumentation,
+  PerformanceInstrumentation,
 } from './instrumentations';
 export type { ConsoleInstrumentationOptions, ErrorEvent, ExtendedPromiseRejectionEvent } from './instrumentations';